requirements.txt
click==7.1.2
# Elasticsearch 7.x
elasticsearch>=7,<8
# Elasticsearch 7.x
elasticsearch-dsl>=7.0.0,<8.0.0

PyYAML~=5.4.1
setuptools~=54.2.0